requirements.txt
pandas
tqdm
mysql-connector-python
faker
sqlalchemy